concurrency/concurrency-course/client/requirements.txt
GitPython>=2.1.8
python-gitlab>=1.3.0
click
termcolor
argcomplete
